+++ internal/config/ci.go
@@ -7,6 +7,7 @@ import (
 
 type CI struct {
 	Include    []string `hcl:"include,optional" json:"include,omitempty"`
+	Exclude    []string `hcl:"exclude,optional" json:"exclude,omitempty"`
 	MaxCommits int      `hcl:"maxCommits,optional" json:"maxCommits,omitempty"`
 	BaseBranch string   `hcl:"baseBranch,optional" json:"baseBranch,omitempty"`
 }
@@ -17,6 +18,13 @@ func (ci CI) validate() error {
 	}
 
 	for _, pattern := range ci.Include {
+		_, err := regexp.Compile(pattern)
+		if err != nil {
+			return err
+		}
+	}
+
+	for _, pattern := range ci.Exclude {
 		_, err := regexp.Compile(pattern)
 		if err != nil {
 			return err